Support top level require
In projects where we load the gem, it'd be nice to be able to require
the entire library, like:

```
require 'cgla_case_chart_assister`
```

To do this, we need to load all of the project files in
`lib/cgla_case_chart_assister.rb`. Since the primary way we intend to
use these classes in host projects currently is in the parser, I
modified the parser specs to only require the library (to confirm our
use case still works through tests).