Convert all structs to classes
Now that we're using the gem in a host application, it looks like it
would be useful to have defaults for certain model attributes (empty
arrays, for example). This is difficult in structs, so this commit
converts all existing structs into classes. It also creates an abstract
Event model from which our Charge and Arrest events can descend (more
closely expresses how we designed them!).